Remove deprecated dojo functions
